refactor(trees): move trees into src/mountains/trees.ts (peer of rocks), kill the cycle
Trees and rocks are both "scatter scenery on the terrain by reading the analytic
samplers", but after the mountains split rocks lived in src/mountains/ while trees
stayed a top-level sibling that imported the Mountains facade — an asymmetry that
also kept the mountains <-> trees circular import alive.

Move the implementation to src/mountains/trees.ts (a peer of rocks.ts) behind a thin
src/trees.ts facade (export * from './mountains/trees.js'), so every ./trees.js
importer (snow.ts, scene-setup.ts, the bundle) keeps resolving a sibling file. The
moved module now reads the terrain samplers from the leaf modules directly
(./terrain.js getTerrainHeight/getTerrainGradient, ./noise.js forestDensityField)
instead of the Mountains object, which **removes the mountains <-> trees circular
import** — it is now a one-way terrain-mesh -> trees -> terrain/noise dependency.
terrain-mesh.ts imports the sibling ./trees.js directly.

Behaviour-preserving (tree math byte-identical): the only logic change is the local
terrain-sampler wrappers now delegate to the imported leaves (the defensive
Mountains-undefined fallbacks are dead under static ESM imports and were dropped).

Guard: tests/contract-surface-tests.js gains a Section G ./trees.js facade seam
(Trees object + method surface), mirroring the ./mountains.js (F) and ./snowman.js
(D) seams.

Verified: tsc + lint clean (27 pre-existing warnings), full Node suite 0 failed
(contract-surface 68/0, tree-collision 3/3), physics-invariant harness OK, verify
49/49, `npm run build` OK.
